+++ R/makeClusterSequential.R
@@ -21,6 +21,12 @@
 #' str(y)
 #' stopifnot(abc == 3.14)
 #'
+#' clusterExport(cl, "abc", envir = environment())
+#' rm(abc)
+#' y <- clusterEvalQ(cl, { abc })
+#' str(y)
+#' stopifnot(y[[1]] == 3.14)
+#'
 #' @details
 #' Expression and function calls are evaluated in a local environment,
 #' inheriting the global environment.
@@ -77,6 +83,15 @@ sendData.sequential_node <- function(node, data) {
     ## Don't evaluate in the global environment, which is the default
     if (identical(args[["envir"]], globalenv())) {
       args[["envir"]] <- envir
+    }
+
+    ## WORKAROUND: parallel::clusterExport() assigns to global environment
+    ## of the worker, but we want it to be our local environment
+    if (identical(fun, environment(parallel::clusterExport)[["gets"]])) {
+      fun <- function(n, v) {
+        assign(n, v, envir = envir)
+        NULL
+      }
     }
     
     success <- TRUE