+++ inst/testme/test-makeClusterSequential.R
@@ -51,6 +51,14 @@ if (getRversion() < "4.4.0") {
   y <- clusterEvalQ(cl, { abc <- 42; abc })
   stopifnot(identical(y, list(42)), abc == 3.14)
 
+  ## Export a variable to the cluster nodes
+  def <- 42
+  y <- clusterExport(cl, "def")
+  rm(def)
+  stopifnot(!exists("def"))
+  y <- clusterEvalQ(cl, { def })
+  stopifnot(identical(y, list(42)))
+
   ## Test with run-time errors
   y <- tryCatch(parLapply(cl, X = 1, fun = stop), error = identity)
   stopifnot(inherits(y, "error"))